Consolidate parsers to parser.hpp
fix #405
