+++ include/beast/http/impl/parser.ipp
@@ -5,11 +5,20 @@
 // file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 //
 
-#ifndef BEAST_HTTP_IMPL_MESSAGE_PARSER_IPP
-#define BEAST_HTTP_IMPL_MESSAGE_PARSER_IPP
+#ifndef BEAST_HTTP_IMPL_PARSER_IPP
+#define BEAST_HTTP_IMPL_PARSER_IPP
 
 namespace beast {
 namespace http {
+
+template<bool isRequest, class Fields>
+template<class Arg0, class... ArgN, class>
+header_parser<isRequest, Fields>::
+header_parser(Arg0&& arg0, ArgN&&... argn)
+    : h_(std::forward<Arg0>(arg0),
+         std::forward<ArgN>(argn)...)
+{
+}
 
 template<bool isRequest, class Body, class Fields>
 template<class Arg1, class... ArgN, class>
